+++ modules/command/common/send-telemetry-command.js
@@ -1,13 +1,13 @@
-const { execSync } = require('child_process');
+const {execSync} = require('child_process');
 const Command = require('../command');
 const axios = require('axios');
 const FormData = require('form-data');
 const fs = require('fs');
 const converter = require('json-2-csv');
 const split = require('split');
-const { finished } = require('stream');
+const {finished} = require('stream');
 // Constructing promisify from util
-const { promisify } = require('util');
+const {promisify} = require('util');
 const path = require("path");
 // Defining finishedAsync method
 const finishedAsync = promisify(finished);
@@ -17,6 +17,7 @@ class SendTelemetryCommand extends Command {
         super(ctx);
         this.logger = ctx.logger;
         this.config = ctx.config;
+        this.publishService = ctx.publishService;
         this.logFilename = 'active.log';
         this.csvFilename = 'telhub_logs.csv'
     }
@@ -26,7 +27,7 @@ class SendTelemetryCommand extends Command {
      * @param command
      */
     async execute(command) {
-        if(!this.config.telemetryHub.enabled) {
+        if (!this.config.telemetryHub.enabled) {
             return Command.empty();
         }
         const otNodeLogsPath = path.join(path.resolve(__dirname, '../../../'), 'logs');
@@ -36,7 +37,7 @@ class SendTelemetryCommand extends Command {
 
         try {
             await execSync(`cat ${path.join(otNodeLogsPath, this.logFilename)} | grep \'"level":15\' | grep -v \'level-change\' > ${intermediateConversionFile}`);
-        } catch(e) {
+        } catch (e) {
             // No data to be send, finish command
             return Command.repeat();
         }
@@ -44,7 +45,7 @@ class SendTelemetryCommand extends Command {
         // Read json objects from log
         let jsonLogObjects = [];
         const readable = fs.createReadStream(intermediateConversionFile)
-            .pipe(split(JSON.parse, null, { trailing: false }))
+            .pipe(split(JSON.parse, null, {trailing: false}))
             .on('data', function (obj) {
                 jsonLogObjects.push(obj);
             })
@@ -53,6 +54,22 @@ class SendTelemetryCommand extends Command {
             });
         await finishedAsync(readable);
 
+        const jsonld = {
+            "@context": "http://schema.org/",
+            "@type": "OTTelemetry",
+            "minTimestamp": Math.min(...jsonLogObjects.map(x=>x.time)),
+            "maxTimestamp": Math.max(...jsonLogObjects.map(x=>x.time)),
+            "data": jsonLogObjects.map(x => ({
+                "eventName": x.Event_name,
+                "eventTimestamp": x.time,
+                "operationId": x.Id_operation,
+                "operationName": x.Operation_name,
+                "msg": x.msg,
+            }))
+        };
+
+        const assertion = await this.publishService.publish(JSON.stringify(jsonld), '.json', [], ['OTTelemetry'], true, null);
+
         // Convert json objects into csv lines and store them
         await converter.json2csv(jsonLogObjects, (err, csv) => {
                 if (err) {
@@ -60,13 +77,15 @@ class SendTelemetryCommand extends Command {
                 }
                 fs.writeFileSync(`${path.join(otNodeLogsPath, this.csvFilename)}`, csv);
             },
-            { keys:
-                [
-                    {field:'hostname', title: 'Id_node'}, 'Id_operation', 'Operation_name',
-                    'Event_name', {field: 'time', title:'Event_time'}, 'Event_value1',
-                    'Event_value2','Event_value3', 'Event_value4'
-                ]
-            , emptyFieldValue: null});
+            {
+                keys:
+                    [
+                        {field: 'hostname', title: 'Id_node'}, 'Id_operation', 'Operation_name',
+                        'Event_name', {field: 'time', title: 'Event_time'}, 'Event_value1',
+                        'Event_value2', 'Event_value3', 'Event_value4'
+                    ]
+                , emptyFieldValue: null
+            });
 
         // Send csv file to telemetry hub
         let data = new FormData();
@@ -78,7 +97,7 @@ class SendTelemetryCommand extends Command {
                 headers: {
                     ...data.getHeaders()
                 },
-                data : data
+                data: data
             });
         } catch (e) {
             this.logger.error(`Error while sending telemetry data to Telemetry hub: ${e}`);
@@ -88,7 +107,7 @@ class SendTelemetryCommand extends Command {
         execSync(`rm ${intermediateConversionFile}`);
         // Make a copy of log file
         const newLogName = `${new Date().toISOString().slice(0, 10)}-${(Math.random() + 1).toString(36).substring(7)}.log`
-        await execSync(`cp ${path.join(otNodeLogsPath, this.logFilename)} ${path.join(otNodeLogsPath,newLogName )}`);
+        await execSync(`cp ${path.join(otNodeLogsPath, this.logFilename)} ${path.join(otNodeLogsPath, newLogName)}`);
         // Truncate log file - leave only last 30 lines
         await execSync(`tail -n30 ${path.join(otNodeLogsPath, this.logFilename)} > ${path.join(otNodeLogsPath, this.logFilename)}`);
 
@@ -107,7 +126,7 @@ class SendTelemetryCommand extends Command {
             data: {
                 message: '',
             },
-            period: 20 * 60 * 1000, // 5 * 60 * 1000
+            period: 1 * 60 * 1000, // 5 * 60 * 1000
             transactional: false,
         };
         Object.assign(command, map);
